Make Fleks multiplatform compatible
- Adapt registering of systems in World

Instantiation of generic classes is not possible in Kotlin
multiplatform. Thus, instead of instantiating a generic class in the
SystemService the system function gets a factory method for calling
the constructor of the system.

- Introduce adding of component and its listener in world configuration

The component listener will be added together with registering the
component in the world configuration. That was done because when
registering the component listener the base type of the component is
needed. But without reflections it is not possible to detect the type
of component listeners. Thus, adding component and listener objects
together is the most elegant workaround - I hope.

- Adapted readme to reflect changed API

- Adapted Fleks benchmark

Currently registering of systems is not working in the benchmark
`fleks.kt`. In Korge it is working that way. It could be that this is
due to some multiplatform configuration in Korge. But currently this is
beyond my understanding. :(