Add Kotlin Multiplatform info in readme
Updated the github workflow URL since it should build the "kmp" branch.
Updated the version information for the KMP release: 1.0-KMP-RC3. I guess it
would be good if the kmp branch would be updated with content of minor releases
on the master branch. Thus keeping the versions equal between master and kmp
branches. Where the kmp branch adds "-KMP" to the version string.