Implement dependency injection by type names
In case of injection of multiple dependencies with the same type it is now
possible to specify the type name on injection to the world configuration.
When passing that dependency in a system or component listener than the
type name can be given as an argument to Inject.dependency() function call.

Updated the readme to reflect that injection of dependencies is possible
by providing an optional type name as parameter together with the depencency
object.