Update with changes done in korge-fleks (#37)
New feature was added which makes it possible to mark an injectable as
"used". Thus Fleks will not throw exception if the injectable is not
used directly at creation of system and listener objects. Ie. that makes
it possible to use the injected objects during init functions. Default
for use is still "false" as before.