Add kotlinx serialization support to fleks
To be able to serialize world snapshot of fleks the kotlinx serializer
needs to know "how" to serialize Entity value class. Thus, it would be
nice to add @Serializable annotation from kotlinx.serialization directly
to the Entity class declaration. With that it is possible to serialize
the world snapshot (or snapshotOf) in a kotlin multiplatform game that
uses kotlinx.serialization.