Smaller bitwise optimizations.
These range from things that might not matter at all (because the compiler uses them behind-the-scenes), to taking advantage of non-obvious but specified behavior of bitwise operators. The former includes using `ushr 6` instead of `/ 64`, which does have an advantage in that negative bit indices will be treated as unsigned int indices, which still fit inside a large-ish `Array` (it needs 4GB of RAM to store 2 to the 26 `Long` items, though). The latter includes removing manual modulus on shift amounts, since according to the Kotlin docs on shl and ushr, among others, shift amounts applied to `Long`s are implicitly masked to only use the bottom 6 bits (when applied to `Int`s, they use 5 bits).