Use functions with intrinsics where possible.
The `countLeadingZeroBits()` and `countOneBits()` functions each are marked as `@IntrinsicCandidate` on the JVM, because they compile down to a much faster-than-normal piece of native code (usually one instruction, with a name like CLZ or POPCNT). Using intrinsics wherever possible in bit-twiddling code can yield really serious performance gains if the code before used many instructions and it can be replaced by one. I believe `countOneBits()` had its performance improve on a very recent Java version, but before that some code needed to avoid it because it didn't play nicely with  some loop optimizations (though in this case, if you wanted to avoid it, there are better options than looping over 64 bits to count the 1s).