The good stuff: forEachSetBit() and toEntityBag().
These are... pretty crazy pieces of code. They make heavy use of a pattern where they look into an item from the values (just a Long, so this is by value), find the highest bit index, do something with it, and then remove that highest bit from the Long value we have locally. Doing this means that if only one bit is set in the Long we look at, we only have to iterate once through that Long. Two bits set means two iterations, and so on.